services/preprocessing/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [ "src/" ]

[project]
name = "preprocessing"
version = "0.1.0"
description = "Service for preprocessing text data"
requires-python = ">=3.13.2"
dependencies = [
    "cachetools>=6.1.0",
    "faker>=37.4.2",
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "opentelemetry-distro[otlp]>=0.55b1",
    "opentelemetry-instrumentation-fastapi>=0.55b1",
    "opentelemetry-instrumentation-requests>=0.55b1",
    "redis>=6.2.0",
    "structlog>=25.3.0",
    "uvicorn>=0.34.2",
]
